Add shim for MediaStreamTrack
Adding listeners for MediaStreamTrack objects and dispatching events has
no effect in Firefox, although it is supported in Chromium. Due to this
a shim that implements events in MediaStreamTrack is added.

Note that the implementation is a simplified one adjusted only to what
will be needed and it does not take into account advanced parameters of
event listeners like "options" or "useCapture".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>