Handle changes in the selected audio or video input devices during calls
When an input device is changed during a call the old audio or video
track is stopped, a new one is requested, and then the old track is
replaced by the new one in the RTCPeerConnection.

When a track is replaced in an RTCPeerConnection sometimes a
renegotiation is needed. As neither the web UI nor the mobile apps
support a proper renegotiation for now a forced reconnection is used.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>