Fix forced reconnections for guests when the HPB is used
When a reconnection is forced and the standalone signaling server is
used the participant joins again the conversation to get a new session,
and then the websocket to the standalone signaling server is closed,
which causes the reconnection with the signaling server. However, when
the participant is a guest and joins again the previous session gets a
"disinvite" message, which causes the browser to be redirected to the
main Talk page.

Due to this, now the "disinvite" message will be ignored when a forced
reconnection is started until the socket has reconnected. It seems that
the "disinvite" message will always be delivered before the request to
rejoin a conversation returns, so the flag could be disabled as soon as
the request returned instead of when the socket is connected. However,
disabling it once the socket has connected again is also logical and
seems more future proof.

In any case, as the "disinvite" message is ignored that means that if a
participant is actually kicked out from the conversation while
performing a forced reconnection the browser will not be aware of it and
a wrong UI would be shown. However, the chance of that happening is
quite low.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>