Dispatch "addtrack" and "removetrack" events in MediaStream shim
Neither Firefox nor Chromium dispatch "addtrack" and "removetrack"
events when tracks are added and removed to and from a MediaStream.
However, this is part of the specification, so it is expected that all
browsers implement that eventually. Due to that the shim checks whether
the events were dispatched or not in order to do it only when needed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>