Stop audio monitors when all audio tracks end
Until now the audio monitors were stopped when the source stream from
which the audio monitor stream was cloned was removed. Now the audio
monitors are stopped as soon as all* the audio tracks in the audio
monitor stream ends, which will make possible to set an audio monitor
again on a previously used stream from which all the audio tracks, but
not the video tracks, were removed.

*As internally the audio monitor uses a MediaStreamAudioSourceNode
actually only the audio track with the lower id will be used. But for
simplicity, and given that only a single audio track is expected anyway,
all the audio tracks in the audio monitor stream are checked.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>