Add and remove tracks in sync with the original on a linked stream
Until now a linked stream stopped its tracks when the original tracks
were stopped. Now linked tracks are added and removed too when tracks
are added and removed to and from the original stream. This will make
possible to keep audio monitor streams in sync with the original ones
without explicit handling.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>