Do not request streams again while waiting for a previous one
When the selected audio or video device changes a stream from that
device is requested to replace the current one. The stream is resolved
asynchronously, so while that happens the user could select a different
device, which in turn will request another stream. As the previous one
would not have been set yet it will not be replaced by the new one, and
instead several streams would end being added.

Now this is enforced by preventing further stream requests while a
previous one has not been completed yet. If several device changes are
triggered while waiting for a previous one once that previous one is
finished a new stream will be requested for the last selected device.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>