Do not use "finally" on promises returned by "getUserMedia"
"Promise.finally" is available in older browsers for "hand-made"
promises, probably due to a shim or to the code being transpiled (I have
not checked). However, it is not available for internal promises
returned by "navigator.mediaDevices.getUserMedia()".

Moreover, in some browser versions (Firefox >= 58 && <= 68) there seems
to be a clash between the internal promise and the shimed/transpiled
one, as "Promise.finally" is available for "getUserMedia" promises, but
not when used in Nextcloud.

Due to all that "finally" uses were replaced by a function called at the
end of "then" and "catch".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>