Ignore "forceMute" message if already muted
Before the UI only allowed to send the "forceMute" message to
participants that were not muted. Now that moderators can mute all the
other participants at once the message may be received by participants
that are already muted, so it needs to be ignored in that case.
Otherwise the user would see a "You have been muted by a moderator"
notification even if already muted.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>