Get rid of "hark" data channel
The Web UI sends "speaking" and "stoppedSpeaking" messages through the
"status" and "hark" data channels, but the messages are simply ignored
when received in the "hark" data channel. The Android and iOS apps do
not even use the "hark" data channel, neither for sending nor for
listening, so it can be fully removed.

Moreover, with latest Janus versions the peers receive all data channel
messages in the "JanusDataChannel" data channel, no matter on which data
channel they were originally sent. As the "speaking" and
"stoppedSpeaking" messages are sent through the "status" data channel
sending them too through the "hark" data channel caused the messages to
be handled twice.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>