Fix getting messages again after the component was destroyed
When getting new messages, no matter if it succeeds or fails, another
request is triggered again after a small delay. When the MessagesList
component is destroyed the current request is cancelled, which would
stop further requests. However, if the MessagesList is destroyed during
that delay there will be no current request yet, so once the timeout
ends the request will start anyway, and thus the request loop will
continue again and again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>