Fix some signaling messages not handled with external signaling server
Besides the session ID the SimpleWebRTC code adds a "sid" property to
identify Peers, and signaling messages sent by those Peers include the
"sid" property. When the standalone signaling server is not used this
property makes possible to know if a message from certain participant is
being sent by a "stale" Peer object (a Peer object that has been
replaced with a new one, for example due to reconnection), in which case
it is ignored.

However, when the standalone signaling server is used the "sid" property
is not relayed in offer messages. In this case, when the remote Peer
object is created an arbitrary "sid" is assigned, which does not match
with the "sid" of the local Peer object. As the signaling messages sent
with the local Peer object have one "sid" but the Peer object for that
participant in the remote end has a different "sid" those signaling
messages ("mute", "unmute" and the currently unused "connectivityError")
were ignored.

Due to this now the "sid" property is taken into account only when the
standalone signaling server is not used.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>