Send initial state messages for connections also through signaling
When "audioOn", "audioOff", "videoOn" or "videoOff" events are triggered
in the SimpleWebRTC object the message is sent to all the participants
through the data channels. When those events are triggered in the WebRTC
object (for example, from the LocalMedia object) "mute" or "unmute"
messages are sent to all the participants through signaling. The
SimpleWebRTC object proxies all the events from its WebRTC object, so
triggering those events in the WebRTC objects causes both data channel
and signaling messages to be sent.

However, when a connection is established with another participant the
events were triggered in the SimpleWebRTC object, so the messages were
sent only through the data channel. For consistency with the rest of
cases, and as signaling messages are preferred over data channel ones,
now when a connection with another Peer is established the events are
triggered on the WebRTC object instead.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>