Remove unneeded code
When the HPB is used now the initial state is repeated several times
after a connection with another participant is established, so the
little wait here is no longer needed. When the HPB is not used the
initial state is send just once, but as in that case both participants
are already aware of each other the little wait is not needed either.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>