Stop sending the nick through data channels after some time
Since the introduction of the HPB the nick has been sent over and over
again every second, as it is not possible to know when other
participants have established a data channel connection.

However, it is possible to know when other participants join the call,
and it is safe to assume that the other participant established a
connection with the local one after a "reasonable" amount of time.

As sending the nick every second is too aggressive (specially in calls
with a lot of users) now the nick is sent when the local participant
establishes a connection with the remote one and, as that does not
guarantee that the remote one has established a connection with the
local one, the nick is sent again several times (with an increasing
interval) in the next 30 seconds. If a connection is established with
another participant in the meantime the cycle starts again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>