Save the query to count the user for moderators per room every 30sec
The idea of this option was to not show the "leave" button when
you are the last "non-guest" user, so you need to delete it instead.
However this "neat" functionality could be out of date info,
if the other user already left in parallel.
So it doesn't qualify for the O(n) queries every 30 seconds.

Signed-off-by: Joas Schilling <coding@schilljs.com>