Remember selected input devices
Before the selected input devices were used as long as the device was
still available and Talk was still open. However, if the device was
disconnected and connected it was not automatically selected again, even
if Talk was still open. Now the last selected device of each kind is
remembered, so it will be automatically used if available when joining a
call, even if it was disconnected and connected or if Talk was closed
and opened.

Additionally, if audio or video devices were explicitly disabled this
will also be remembered and respected when Talk is open again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>