Do not remove ended streams if there are pending changes
When all the tracks from a stream were ended the stream was immediately
removed. However, if there were pending audio or video changes the
stream could be reused to add new tracks once "getUserMedia" finishes
without having to create a new one.

Moreover, this also fixes an issue when changing audio and video devices
at the "same" time. When a device is changed first the tracks for the
previous device are stopped. If both the audio and video devices were
changed then all the tracks of the stream were ended, and thus the
stream was removed. However, the stream was still referenced in the
"localStreamsChanged" arrays of the handlers, so when "getUserMedia"
returned the new track the old stream was not found, a new one was
created and the track was added to it. As this happened for both audio
and video two different streams ended being created, each one only with
audio tracks or video tracks, and as both streams were passed to
"localStreamChanged" either the audio or the video, depending on the
last stream passed, were disabled due to the stream not having tracks of
that type.

Something similar happened too when "getUserMedia" failed, although in
that case the old stream was the one passed to "localStreamChanged".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>