Mark handling of input id changes as pending since the beginning
Before the audio and video input id changes were marked as pending only
while waiting for "getUserMedia" to finish. However, in the initial
phase of handling input id changes the tracks for the previous device
are stopped, which could cause the stream to be removed (for example, if
there was just a single active device) even if it is still referenced in
the "localStreamsChanged" array. Now handling input id changes are
marked as pending since the beginning, which prevents ended streams to
be removed while they may still be needed later.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>