Wait for pending devices enumeration before getting user media
When the devices are enumerated the selected audio and video input id
may change depending on which devices are available. Therefore, when
"getUserMedia" is called if an enumeration has been triggered but it has
not finished yet now it will be waited for the enumeration to finish
before getting the user media.

Enumerating the devices is a fast operation, so this should not cause
any noticeable delay when getting the user media.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>