Ensure that devices were enumerated when starting the local media
When starting the local media for the first time the media devices could
have not been enumerated yet, so even if the remembered devices are
available they were not used in the first "getUserMedia" request.
However, as the devices are enumerated after getting the user media (as
that is the only time when it is guaranteed that the full device
information can be got) this caused the remembered devices to be then
found and further "getUserMedia" requests to be made for them.

To avoid an initial "getUserMedia" request with default ids followed by
other requests with the remembered ones now the media devices are tried
to be enumerated before starting a call.

This does not guarantee that the extra "getUserMedia" requests will be
avoided, as if persistent media permissions are not granted the browser
may not provide the full device IDs, but at least avoids them in some
cases.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>