Adjust hacked tabs style to changes in nextcloud-vue 2.4.0
Since nextcloud-vue 2.4.0 the tabs content container is no longer a
direct child of the app sidebar and it is now wrapped in an
".app-sidebar-tabs" container. Therefore the hacked minimum height and
overflow need to be set in that container instead (although the hidden
overflow still needs to be propagated to the tabs content container).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>