Center the empty content in the tab instead of just adding a margin
Since nextcloud-vue 2.0.0 the server styling is not used for the
sidebar. Therefore the reduced top margin (10vh) for empty contents in
"#app-sidebar" is no longer applied. As the default top margin is used,
which is designed for the main view and thus is quite large (30vh), the
empty content in the chat tab appears very low. Moreover, no sidebar is
shown if the window height is small, so the button to join the
conversation may not be reachable. Instead now the empty content is
centered in the tab contents using a flex display, which also causes a
sidebar to be shown if there is not enough available space to show the
contents.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>