Fix sending initial media state to participants without audio or video
When the HPB is used if a remote participant has no audio or video no
Peer will be created for that participant. Until now no initial media
state was sent for those participants, so a wrong initial media state of
the local participant could be seen by those remote participants. Now
the current media state is sent to those participants when they join the
conversation.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>