Stop checking for peer media if audio/video status is received
As until now the initial media state message was sent just once and it
could be missed the received media was locally checked to enable the
media if needed. Unfortunately those checks were not as reliable as
desirable, so now they are stopped as soon as a media state message from
from the other peer is received. This implicitly fixes things like
never stopping checking the video if the other peer has no camera
enabled.

The checks can not be fully removed yet, as the repeated sending of
initial media state is currently implemented only in the WebUI, but not
in the mobile apps.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>