Unify sending "nickChanged" data channel messages
Instead of sending "nickChanged" data channel messages from different
places now an internal "nickChanged" event with the name is emitted, and
its handler takes care of sending the data channel message as needed.

This also fixes trying to send data channel messages to receiving only
peers when the nick is changed and the HPB is used.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>