Remove "userId" property from internal "nick" event
The "userId" property was used only in CallParticipantModel, but as it
is also set from the participant information provided by the signaling
in most cases it is not needed.

The only case in which it would be needed is if the Peer object is
created and connected before the participant is found in the participant
information provided by the signaling. However even in that rare case
showing a user as a guest would happen very briefly and it will correct
itself as soon as the signaling sends the participant information. Due
to this, and as it will simplify also a "nickChanged" signaling message
to be introduced later, the "userId" property is now ignored.

Note that this only affects the internal "nick" event; the "userId"
property still needs to be sent in the "nickChanged" data channel
message as it is used by the Android app.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>