Adjust Files sidebar tab to changes in server
When a tab is registered in the Files sidebar a component is no longer
provided. Instead an object with some attributes that describe
the tab (like id or icon) and some callbacks for lifecycle events is
expected now.

Essentially the AppSidebarTab component is now created by the Sidebar
itself instead of being given it. However, as the component needed some
adjustments in its CSS style that needs to be forced when the child
component is mounted.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>