In case of error when sending message, revert
Whenever an error occurs when sending a message, the message is now put
back into the field and the temporary one is removed from the messages list.

Whenever a conversation was locked or lobby was enabled concurrently,
posting a message before the UI update could result in a 403 or 412 error.
These now display a proper message to inform about missing permissions.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>