Fix empty call view position in sidebar
When the call view is shown in the Files app or public share page
sidebar its direct children (the empty call view and the wrapper of
videos) are expected to have an absolute position. However, as the empty
call view had a relative position it was prepended before the call view
padding (which defines the call view height) instead of fully covering
it.

When the call view is shown in the main Talk UI or in the video
verification sidebar the call view has an explicit height attribute
(rather than using a padding), so in this case it does not matter if the
empty call view has a relative or absolute position.

However, when the empty call view is shown in the grid it must use a
relative position.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>