Chat previews now expanded with aspect ratio
Instead of cropping the previews in the conversation to a square, they
are now using a limited height but their width is computed based on
aspect ratio. This is done by passing different arguments to the
previews endpoint so it delivers a preview with the matching size.

Adjusted styles around the preview to make it look better.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>