Rework scroll into view of conversation after post
Fixes an issue where a chat in the files sidebar would trigger a warning
because the conversation list element don't exist there.

When posting a message, an event is now triggered so other components
could react on it.

The conversation list will now take the responsibility of scrolling the
current conversation into view after posting, instead of the
NewMessageForm.

The behavior is also adjusted to always scroll regardless where the
conversation is located in the list.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>
Co-Authored-by: Daniel Calviño Sánchez <danxuliu@gmail.com>