Move the filename of attachments to a tooltip
Instead of displaying file names below attachments in a conversations,
like for images, it is now displayed only on hover in a tooltip.

Howver, in the upload editor the file names are still shown as they can
be helpful in that situation.

This cleans up the message list a bit and also solves some layout issues
with images in narrow views like the sidebar.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>