Clear state when starting or discarding uploads
The upload editor is shown based on both "showUploadEditor" (set in the
store) and "modalDismissed" (set in the component). As "modalDismissed"
is initially false and "showUploadEditor" was not cleared after an
upload started or was discarded this caused the upload editor to be
immediately shown if mounted again (for example, when joining a
conversation in the Files app sidebar) if a previous upload was
initialised.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>