Highlight quote after scrolling to it
When clicking on a quote, we scroll to it and now there's a new
highlight animation that fades out

The animation and scrolling logic was moved from the Quote vue to the
Message vue. The Quote vue will now only send an event to tell that we
want to focus a specific message.

A DOM-level method was used for triggering the highlight to be able to
access it with document.getElementById() as it's cheaper than trying to
look up the message by id within the collection of MessagesList.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>