Highlight row initially when message in URL hash
When coming from the unified search or other places, the URL hash will
contain the id of a message. The message in question will now be
highlighted for a short time.

Increased animation from 2 to 5 seconds to make it easier to spot when
coming from the URL/anchor.

Changed message highlight color to be the existing hover one.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>