Don't leave conversation on URL hash change
Whenever only the URL hash changes despite the token staying the same, a
route change event is still triggered.

This fix avoids leaving and re-joining the same conversation.

Fixes an issue where clicking a search result would trigger both the
event that leaves the conversation within a route change and a page
unload event, which also leaves the same conversation.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>
Co-authored-by: Joas Schilling <coding@schilljs.com>