Scroll to center, even after coming from search result
When clicking the search result, the native anchor logic makes the
browser scroll directly to the message, but leaves it at the top of the
window. To be able to force scrollIntoView, we needed to add some delay
and scroll afterwards a bit further.

The scrolling is done to the center of the messages list.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>