Return raw description also when getting the conversation list
As the description length is limited and the response for the
conversation list endpoint is gzip compressed the hash does not provide
too much of a reduction, but introduces a more complex logic in the
clients. Due to all this, for simplicity now the raw description is also
returned when getting the conversation list.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>