Fix additional participants being able to join a password request room
This fixes a regression introduced in 4afa2d7946.

The active guests are only those who are currently in a call, and not
those who are currently in the conversation. Therefore other guests or
users were not prevented from joining a password request conversation if
a guest was in the conversation but not in the call.

Fortunately in practice this was not a problem, as the Web UI starts a
call immediately after joining the conversation, which made the guest
immediately active and thus prevented others from joining.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>