Fix owner being able to add more users to a password request room
Only the owner and another participant will be allowed to join a
password request room, so there is no point in being able to add more
participants to those rooms.

Although throwing the exception in the listener is enough to prevent
adding the participants unhandled exceptions in the endpoint are
returned as error 404, but the expected error would be 400. To minimize
conflicts with other pull requests and backports it is explicitly
checked if the room is a password request room instead of refactoring
the code to handle the exception.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>