Replace blur filter with a blurred canvas image
The CSS blur filter in the video background causes Chromium's UI to be
very slow. The performance hit is larger when the element to blur is
larger, and also when the blur radius is larger, so a promoted speaker
with a blurred image seems to be the worst case (even worse than a grid
with several rows and columns).

However, even if the CSS blur filter is slow, applying the blur filter
on a canvas is much more performant (maybe the blur filter itself has
the same performance in both cases, but the UI uses it several times in
a row when rendering the element, but that is just a guess).

Due to this now a blurred background image is generated from the
original background image whenever the blur size changes, and that
blurred image is the one used in the image element without any CSS
filter.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>