Remove unneeded code
Neither "webrtcSupport.prefix === 'moz'", "pcConfig.iceTransports" or
"candidate.candidate.candidate" were ever set, so the check always
failed.

Moreover, currently candidates are restricted based on the transport
only to test the TURN server. This is done through the
"iceTransportRelay" parameter and is handled by the browser itself, so
there is no need to fix the code, it can be simply removed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>