Fix not sending signaling messages to participants without a Peer object
"webrtc.sendToAll" only sends the signaling message to participants for
which there is a Peer object. Due to this now it also emits an event
that is handled by the WebRTC wrapper to send the message too to the
rest of participants in the call that do not have a Peer object.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>