Do not fetch the source image again to generate the blurred one
Until now a new source image was loaded every time that the blurred
image had to be generated. Now the source image is loaded only when the
URL of the background image changes and that "cached" image is reused to
generate any blurred image.

The source image is loaded using a watcher instead of a computed
property because the "complete" read-only property of HTMLImageElement
is not reactive. Due to this it is not possible to set
"blurredBackgroundImageSource" as a computed property and then check its
"complete" attribute in "generatedBackgroundBlur" to know if the image
is loaded or not.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>