Cache generated blurred images
The cache id needs to be generated from the background image URL rather
than the image itself (as checksuming the image would require drawing it
to a canvas, which although still faster than drawing a blurred image
would defeat the use of a cache).

For simplicity the images are cached locally in the VideoBackground
element rather than in a central store. Due to this the images will be
regenerated again if the component is created again (for example, when
changing between speaker mode or grid view), but minimizes the
posibility of using a wrong cached blurred image if the image returned
by the URL changed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>