Improve room display name protection
In case the API would return a room for some reason, it would return the
name as "Private Conversation". The logic has been updated to also work
properly with listable vs non-listable rooms in regards to guests.

Refactored the listable check logic into a single method to make it
reusable.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>