Use a LRU cache for the blurred background images
The blurred background images were cached directly in an array, so the
cache could grow without limits. Although it should be a rare case, this
could lead to a high memory consumption if the user slowly resizes
during a long time the browser window. Therefore the cache now uses a
least recently used cache limited to just five elements (so it can store
the blurred backgrounds when the sidebars are open or closed, plus
having a bit of room).

"node-lru-cache" was selected as the library to use due to being small,
widely used and, apparently, matching the expected behaviour
("quick-lru", which is smaller and also widely used, although not so
much, unfortunately does not seem to honour the "maximum size"
parameter).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>