Keep going if loading the image blurrer worker fails
Until now it was assumed that loading the image blurrer worker would not
fail. However, if the browser does not support the CSP "worker-src"
directive (like Safari) trying to load the worker would throw an
exception which would interrupt the load of Talk itself. Now the
exception is caught and logged instead, and the image blurrer reverts to
a synced generation if no worker is available (although that should not
happen, as the image blurrer is used only on Chromium and Chromium
supports the "worker-src" directive).

Note that Firefox does not throw an exception when loading a worker if
not allowed by the CSP. Moreover, as the worker object is created, it
does not seem to be possible to know if the worker is actually not
available. This is not a problem, though; the image blurrer is not used
on Firefox, and even if it was Firefox supports the "worker-src"
directive.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>