Use lazy load for image blurrer worker
The image blurrer worker now is loaded only when it is actually needed,
so this should prevent the worker to be loaded on browsers other than
Chromium, as well as on Chromium if no call is joined.

This indirectly fixes the CSP error shown in Safari when the worker was
loaded (as the worker is no longer loaded now in Safari) due to Safari
not supporting the "worker-src" directive but requiring "child-src" to
be used instead.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>