Get avatar image for non-icon room avatars
The "custom" room avatars are the conversation picture set by the room
participants, so they need to be got from the avatar endpoint. As it is
an OCS endpoint the "OCS-APIRequest" header must be set, so the endpoint
can not be directly used as the URL of a CSS image. Instead a request
with the right headers is performed to the OCS endpoint, the response is
treated as a blob, and then a URL to that blob is used.

The "user" room avatars (those for one-to-one conversations) can be got
in the same way, so othe avatar component is dropped in favour of the
room avatar endpoint.

Besides being more consistent it also removes the user menu that was
shown on the avatar component and felt out of place in the room list
(and also did not properly work, as it was not possible to show the menu
without also changing to the one-to-one conversation).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>