Fix grid view jumping back to first page in some updates
When the number of participants changed or the window was resized the
current page was reset to the first one. However, now that
"displayedVideos" is a computed property and no longer reset directly in
"makeGrid()" it is no longer necessary to go back to the first page.

Moreover, as "numberOfPages" is also a computed property it can be
watched to restrict the current page to be always in range.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>