Fix layout hint not removed in some cases
When there was no longer an overflow the layout hint event was not
triggered with a false value, so even if the window was resized to
accomodate the larger number of participants the hint was not
automatically removed. Now the hint is removed whenever there is no
longer an overflow or the grid is shown as a stripe.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>