Fix "user has avatar" flag set as false for default user avatars
The avatar endpoint returns a picture for both default and custom
avatars, and the Avatar component does not differentiate between them;
as long as a picture is returned by the endpoint the "user has avatar"
flag is set to true.

However, the VideoBackground component only considered an avatar as
having a picture if it was a custom picture, so it would not show the
picture if it was not previously loaded (which should not happen in
practice, as Avatar components were loaded before the call was started).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>