Show the average color of the user avatar as video background in Safari
Safari suffers the same performance hit as Chromium when using the blur
filter. However, Safari does not support using the blur filter on
canvases, so the same workaround used for Chromium does not work in
Safari. To solve this, and to still use a background with colors related
to the avatar, the average of the colors in the avatar is used.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>