Use a shared cache for average colors
Before the average color for video backgrounds was locally stored in
each VideoBackground instance. During a call different VideoBackground
instances can be mounted and destroyed for the same participant (for
example, when switching between speaker mode and grid view, or when a
promoted speaker changes), so the previously calculated value was lost
and the average color had to be recalculated every time that happened.

Now the average colors are globally cached and shared between
VideoBackground instances, so previously calculated values can be used
by new VideoBackground instances.

The global cache is cleared each time a call ends.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>