Fix lobby time input
Lobby timer can now be cleared. When doing so it is saved directly.
Adjusted minute selection to have an interval of 5.
Adjusted default date to be the start of the next hour.
Removed extra click for confirmation as we anyway have the save button.
Changed the data type for communicating with date picker to use
timestamps, since this is what we store anyway.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>