Add config setting to make grid videos limit a hard limit
The hard limit can be enabled with OCC with "occ config:app:set spreed
--value=yes grid_videos_limit_enforced". The changes will take effect
once the user reloads the page

The grid videos limit by default is a fuzzy limit, as the number of
videos shown could be larger due to the available space and aspect ratio
to distribute the videos on. When the limit is made a hard limit the
distributed slots will be the same as before, but only as many videos as
allowed by the limit will be filled in each page.

However, note that with the hard limit the grid may look "broken", as
there is still space to be filled with videos yet they are in the next
page.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>