Add padding for local video in last grid page
When the main grid is paginated the last page has less videos than the
others. As the local video is shown after the remote videos this caused
the local video to be shown at a different position than the rest of the
pages. To solve this now empty divs are added to the last page of the
grid as padding to fill the empty space in the grid and thus place the
local video at the same position as in the other pages.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>