Fix new conversation dialog when no item is selected
"showModal()" was being called as the handler for the on-click event and
the EventBus.$on('NewGroupConversationDialog'). As the first one always
provides the event as the first parameter the item was always defined,
but it never contained a label, so the model became undefined and thus
the dialog failed to open.

Now the method was split in two, one to be used to just show the modal
and another one that also takes into account the item being given.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>