Update last read message marker on scroll
Moved marker behavior from server to client by passing setReadMarker=0
to queries.

Added scroll handler that detects whether the read marker was seen and
scrolled past, if yes it will set the read marker to the first message
visible at the top.

Keep the marker when the document is not focussed.

When opening a conversation, scrolls to the read marker.

The lastMessage attribute of a conversation is now updated earlier to
make sure the read marker can update properly when there are new
messages appended.

Instead of waiting for the next refresh cycle, this fix immediately
updates lastMessage and lastReadMessage after posting a message.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>