Update Room object last message after updating
When updating the last message in the database, we now also update it in
the Room object like it is done for other properties.

Fixes an issue when creating a room with participants that the latter
received a last_read_message id of zero. This is because the code path for
creation is done within a single API call, and while the room got
created, the last_message id in Room was still 0 when read out for
setting to the participants in addUsers().

Signed-off-by: Vincent Petry <vincent@nextcloud.com>