Fix not joining call again on reconnection with HPB
When the connection between a client and the external signaling server
is interrupted the signaling server may unregister its session, even if
the Nextcloud session is still active. Once the client reconnects it
will join the room again, which will generate a new signaling session
for it.

If the connection between the browser and the external signaling server
is interrupted during a call the browser will automatically rejoin the
call after joining the conversation again. However, it will still use
the same Nextcloud session, and that Nextcloud session may still be seen
as "in the call" if the signaling server has not closed yet the expired
session (which is different from unregistering it).

Due to this it must be possible to set the "in call" flags again to the
same value, as even if it does not change the Nextcloud session status
it may change the signaling session status.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>