Fix avatar and video not dimmed when remote participant loses connection
The CSS classes were defined in CallView, but as ".videoContainer" was
set in a subcomponent due to the placement of "::v-deep" in the rule it
had no effect. This was fixed by moving the rule to the subcomponent
instead to get rid of "::v-deep".

Besides that now the full avatar container (that is, including the
coloured background) is dimmed instead of just the avatar itself, as it
makes it more consistent with the dimmed video.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>