Show loading icon on videos too when disconnected
It seems that it is not possible to show a loading icon directly on the
video element (or, at least, I have not found how :-) ), so the video is
wrapped in another element on which the loading icon is applied.

The loading icon on videos is larger than the normal one, as otherwise
it was not noticeable due to being "framed" by the whole video (as
opposed to the normal loading icon on avatars, which is "framed" by the
smaller avatar).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>