Don't redirect to not-found page if user switched away
Whenever a request is running in the background and might detect
that the current conversation or session is gone, it would prompt the
app to redirect to the not found page.

In some cases, the user already switched away to another page while the
request was running. The fix detects such situations and skips the not
found page redirect.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>