Show error notification also when signaling hello message fails
Until now an error notification was shown if the connection of the
websocket to the external signaling server failed. However, if the
websocket connection succeeded but then there was an error when actually
connecting to the signaling server at an application level (the hello
message response was an error) no error notification was shown in the
UI. Now the error notification is also shown in that case, and it will
not be removed (no matter if originally it happened due to an error in
the websocket or in the connection) until the connection has succeeded.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>