Add hint of issue in notification about error connecting to signaling
Several errors in a row in the hello response usually point to a
misconfigured signaling server that can not connect to the Nextcloud
server. The error message now also takes into account the number of
hello response errors in a row and hints at a possible misconfiguration.

As this is a very specific issue the UI does not show any other detail
regarding the received error, and the message is not reverted to the
generic one if the websocket connection fails after the hello response
failed; it is just a hint to make an admin look at the browser console
or the signaling server logs.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>