Keep pending input message per conversation
Save the message from the current input field by conversation into the
store.

This makes it possible to keep the message when switching the
conversation.

Also fixes an issue where joining or leaving a call would clear the
message due to the field being destroyed and recreated between the
message list and the sidebar.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>